+++ integrations/batcher_test.go
@@ -337,9 +337,7 @@ batch
 		.period(10s)
 		.every(10s)
 		.groupBy('cpu')
-	|log().prefix('QUERY')
 	|where(lambda: "mean" < 10)
-	|log().prefix('WHERE')
 	|count('mean')
 	|httpOut('TestBatch_CountEmptyBatch')
 `
@@ -391,9 +389,7 @@ batch
 		.period(10s)
 		.every(10s)
 		.groupBy('cpu')
-	|log().prefix('QUERY')
 	|where(lambda: "mean" < 10)
-	|log().prefix('WHERE')
 	|sum('mean')
 	|httpOut('TestBatch_CountEmptyBatch')
 `
@@ -524,6 +520,45 @@ batch
 
 	testBatcherWithOutput(t, "TestBatch_Default", script, 30*time.Second, er, false)
 }
+func TestBatch_Delete(t *testing.T) {
+
+	var script = `
+batch
+	|query('''
+		SELECT mean("value")
+		FROM "telegraf"."default".cpu_usage_idle
+		WHERE "host" = 'serverA' AND "cpu" = 'cpu-total'
+''')
+		.period(10s)
+		.every(10s)
+		.groupBy(time(2s))
+	|delete()
+		.field('mean')
+		.tag('dc')
+	|default()
+		.field('mean', 10.0)
+		.tag('dc', 'sfc')
+	|groupBy('dc')
+	|sum('mean')
+	|httpOut('TestBatch_Delete')
+`
+
+	er := kapacitor.Result{
+		Series: imodels.Rows{
+			{
+				Name:    "cpu_usage_idle",
+				Tags:    map[string]string{"dc": "sfc"},
+				Columns: []string{"time", "sum"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 8, 0, time.UTC),
+					50.0,
+				}},
+			},
+		},
+	}
+
+	testBatcherWithOutput(t, "TestBatch_Delete", script, 30*time.Second, er, false)
+}
 
 func TestBatch_DoubleGroupBy(t *testing.T) {
 
@@ -557,6 +592,164 @@ batch
 	}
 
 	testBatcherWithOutput(t, "TestBatch_SimpleMR", script, 30*time.Second, er, false)
+}
+
+func TestBatch_GroupByMeasurement(t *testing.T) {
+
+	var script = `
+batch
+	|query('''
+		SELECT mean("value")
+		FROM "telegraf"."default"./cpu_.*/
+		WHERE "host" = 'serverA'
+''')
+		.period(10s)
+		.every(10s)
+		.groupBy(time(2s), 'cpu')
+		.groupByMeasurement()
+	|max('mean')
+	|httpOut('TestBatch_GroupByMeasurement')
+`
+
+	er := kapacitor.Result{
+		Series: imodels.Rows{
+			{
+				Name:    "cpu_usage_user",
+				Tags:    map[string]string{"cpu": "cpu-total"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 28, 0, time.UTC),
+					8.97243107764031,
+				}},
+			},
+			{
+				Name:    "cpu_usage_user",
+				Tags:    map[string]string{"cpu": "cpu0"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 28, 0, time.UTC),
+					8.00000000002001,
+				}},
+			},
+			{
+				Name:    "cpu_usage_user",
+				Tags:    map[string]string{"cpu": "cpu1"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 28, 0, time.UTC),
+					6.49999999996908,
+				}},
+			},
+			{
+				Name:    "cpu_usage_idle",
+				Tags:    map[string]string{"cpu": "cpu-total"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 28, 0, time.UTC),
+					91.06416290101595,
+				}},
+			},
+			{
+				Name:    "cpu_usage_idle",
+				Tags:    map[string]string{"cpu": "cpu0"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 28, 0, time.UTC),
+					85.08910891088406,
+				}},
+			},
+			{
+				Name:    "cpu_usage_idle",
+				Tags:    map[string]string{"cpu": "cpu1"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 28, 0, time.UTC),
+					96.49999999996908,
+				}},
+			},
+		},
+	}
+
+	testBatcherWithOutput(t, "TestBatch_GroupByMeasurement", script, 30*time.Second, er, true)
+}
+func TestBatch_GroupByNodeByMeasurement(t *testing.T) {
+
+	var script = `
+batch
+	|query('''
+		SELECT mean("value")
+		FROM "telegraf"."default"./cpu_.*/
+		WHERE "host" = 'serverA'
+''')
+		.period(10s)
+		.every(10s)
+		.groupBy(time(2s), 'cpu')
+	|groupBy('cpu')
+		.byMeasurement()
+	|max('mean')
+	|httpOut('TestBatch_GroupByMeasurement')
+`
+
+	er := kapacitor.Result{
+		Series: imodels.Rows{
+			{
+				Name:    "cpu_usage_user",
+				Tags:    map[string]string{"cpu": "cpu-total"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 18, 0, time.UTC),
+					9.90919811320221,
+				}},
+			},
+			{
+				Name:    "cpu_usage_user",
+				Tags:    map[string]string{"cpu": "cpu0"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 18, 0, time.UTC),
+					5.93434343435388,
+				}},
+			},
+			{
+				Name:    "cpu_usage_user",
+				Tags:    map[string]string{"cpu": "cpu1"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 18, 0, time.UTC),
+					6.54015887023496,
+				}},
+			},
+			{
+				Name:    "cpu_usage_idle",
+				Tags:    map[string]string{"cpu": "cpu-total"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 18, 0, time.UTC),
+					91.01699558842134,
+				}},
+			},
+			{
+				Name:    "cpu_usage_idle",
+				Tags:    map[string]string{"cpu": "cpu0"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 18, 0, time.UTC),
+					85.93434343435388,
+				}},
+			},
+			{
+				Name:    "cpu_usage_idle",
+				Tags:    map[string]string{"cpu": "cpu1"},
+				Columns: []string{"time", "max"},
+				Values: [][]interface{}{[]interface{}{
+					time.Date(1971, 1, 1, 0, 0, 18, 0, time.UTC),
+					95.98484848485191,
+				}},
+			},
+		},
+	}
+
+	testBatcherWithOutput(t, "TestBatch_GroupByMeasurement", script, 30*time.Second, er, true)
 }
 
 func TestBatch_AlertAll(t *testing.T) {
@@ -665,31 +858,31 @@ batch
 				Values: [][]interface{}{
 					{
 						time.Date(1971, 1, 1, 0, 0, 20, 0, time.UTC),
-						"cpu_usage_idle:cpu=cpu1,",
+						"cpu_usage_idle:cpu=cpu1",
 						"CRITICAL",
 						96.49999999996908,
 					},
 					{
 						time.Date(1971, 1, 1, 0, 0, 22, 0, time.UTC),
-						"cpu_usage_idle:cpu=cpu1,",
+						"cpu_usage_idle:cpu=cpu1",
 						"CRITICAL",
 						93.46464646468584,
 					},
 					{
 						time.Date(1971, 1, 1, 0, 0, 24, 0, time.UTC),
-						"cpu_usage_idle:cpu=cpu1,",
+						"cpu_usage_idle:cpu=cpu1",
 						"CRITICAL",
 						95.00950095007724,
 					},
 					{
 						time.Date(1971, 1, 1, 0, 0, 26, 0, time.UTC),
-						"cpu_usage_idle:cpu=cpu1,",
+						"cpu_usage_idle:cpu=cpu1",
 						"CRITICAL",
 						92.99999999998636,
 					},
 					{
 						time.Date(1971, 1, 1, 0, 0, 28, 0, time.UTC),
-						"cpu_usage_idle:cpu=cpu1,",
+						"cpu_usage_idle:cpu=cpu1",
 						"CRITICAL",
 						90.99999999998545,
 					},
@@ -724,7 +917,7 @@ batch
 		Series: imodels.Rows{
 			{
 				Name:    "cpu_usage_idle",
-				Tags:    map[string]string{"cpu": "cpu1", "level": "CRITICAL", "id": "cpu_usage_idle:cpu=cpu1,"},
+				Tags:    map[string]string{"cpu": "cpu1", "level": "CRITICAL", "id": "cpu_usage_idle:cpu=cpu1"},
 				Columns: []string{"time", "mean"},
 				Values: [][]interface{}{
 					{
@@ -825,8 +1018,8 @@ func TestBatch_AlertStateChangesOnly(t *testing.T) {
 		atomic.AddInt32(&requestCount, 1)
 		if rc := atomic.LoadInt32(&requestCount); rc == 1 {
 			expAd := kapacitor.AlertData{
-				ID:      "cpu_usage_idle:cpu=cpu-total,",
-				Message: "cpu_usage_idle:cpu=cpu-total, is CRITICAL",
+				ID:      "cpu_usage_idle:cpu=cpu-total",
+				Message: "cpu_usage_idle:cpu=cpu-total is CRITICAL",
 				Time:    time.Date(1971, 1, 1, 0, 0, 0, 0, time.UTC),
 				Level:   kapacitor.CritAlert,
 			}
@@ -836,8 +1029,8 @@ func TestBatch_AlertStateChangesOnly(t *testing.T) {
 			}
 		} else {
 			expAd := kapacitor.AlertData{
-				ID:       "cpu_usage_idle:cpu=cpu-total,",
-				Message:  "cpu_usage_idle:cpu=cpu-total, is OK",
+				ID:       "cpu_usage_idle:cpu=cpu-total",
+				Message:  "cpu_usage_idle:cpu=cpu-total is OK",
 				Time:     time.Date(1971, 1, 1, 0, 0, 38, 0, time.UTC),
 				Duration: 38 * time.Second,
 				Level:    kapacitor.OKAlert,
@@ -894,16 +1087,16 @@ func TestBatch_AlertStateChangesOnlyExpired(t *testing.T) {
 		rc := atomic.LoadInt32(&requestCount)
 		if rc < 3 {
 			expAd = kapacitor.AlertData{
-				ID:       "cpu_usage_idle:cpu=cpu-total,",
-				Message:  "cpu_usage_idle:cpu=cpu-total, is CRITICAL",
+				ID:       "cpu_usage_idle:cpu=cpu-total",
+				Message:  "cpu_usage_idle:cpu=cpu-total is CRITICAL",
 				Time:     time.Date(1971, 1, 1, 0, 0, int(rc-1)*20, 0, time.UTC),
 				Duration: time.Duration(rc-1) * 20 * time.Second,
 				Level:    kapacitor.CritAlert,
 			}
 		} else {
 			expAd = kapacitor.AlertData{
-				ID:       "cpu_usage_idle:cpu=cpu-total,",
-				Message:  "cpu_usage_idle:cpu=cpu-total, is OK",
+				ID:       "cpu_usage_idle:cpu=cpu-total",
+				Message:  "cpu_usage_idle:cpu=cpu-total is OK",
 				Time:     time.Date(1971, 1, 1, 0, 0, 38, 0, time.UTC),
 				Duration: 38 * time.Second,
 				Level:    kapacitor.OKAlert,